split electron app into multiple files
