+++ src/electron/env.js
@@ -1,0 +1,5 @@
+const isDev = process.env.NODE_ENV === 'development';
+
+module.exports = {
+  isDev,
+};
